Se elimino la interfaz y clase abstracta para item para mayor simplicidad
