+++ src/main/java/cl/uchile/dcc/finalreality/model/weapon/Weapon.java
@@ -3,13 +3,23 @@ package cl.uchile.dcc.finalreality.model.weapon;
 import cl.uchile.dcc.finalreality.model.character.GameCharacter;
 
 /**
- * This represents a Weapon of the game, a Weapon is also an {@link Item} and
- * it must extend that interface.
+ * This represents a Weapon of the game. A Weapon is contained in the player's inventory.
  * A Weapon can be equipped only by a PlayerCharacter.
  */
-public interface Weapon extends Item {
+public interface Weapon {
   /**
    * Returns the Weapon's damage.
    */
+
   int getDamage();
+
+  /**
+   * Returns the Weapon's name.
+   */
+  String getName();
+
+  /**
+   * Returns the Weapon´s weight.
+   */
+  int getWeight();
 }
